Move /zones/ explicitly to /zones/forward/ and /zones/reverse/.
Resolves #340.
