+++ mreg/api/v1/urls.py
@@ -50,12 +50,17 @@ urlpatterns = [
     re_path(r'^networks/(?P<network>[^/]+/\d+)/unused_list', views.network_unused_list),
     path('txts/', views.TxtList.as_view()),
     path('txts/<pk>', views.TxtDetail.as_view()),
-    path('zones/', views_zones.ZoneList.as_view()),
-    path('zones/hostname/<hostname>', views_zones.zone_by_hostname),
-    re_path(r'^zones/(?P<name>(\d+/)?[^/]+)$', views_zones.ZoneDetail.as_view()),
-    re_path(r'^zones/(?P<name>(\d+/)?[^/]+)/delegations/$', views_zones.ZoneDelegationList.as_view()),
-    re_path(r'^zones/(?P<name>(\d+/)?[^/]+)/delegations/(?P<delegation>(.*))', views_zones.ZoneDelegationDetail.as_view()),
-    re_path(r'^zones/(?P<name>(\d+/)?[^/]+)/nameservers$', views_zones.ZoneNameServerDetail.as_view()),
+    path('zones/forward/', views_zones.ForwardZoneList.as_view()),
+    path('zones/forward/hostname/<hostname>', views_zones.forward_zone_by_hostname),
+    re_path(r'^zones/forward/(?P<name>(\d+/)?[^/]+)$', views_zones.ForwardZoneDetail.as_view()),
+    re_path(r'^zones/forward/(?P<name>(\d+/)?[^/]+)/delegations/$', views_zones.ForwardZoneDelegationList.as_view()),
+    re_path(r'^zones/forward/(?P<name>(\d+/)?[^/]+)/delegations/(?P<delegation>(.*))', views_zones.ForwardZoneDelegationDetail.as_view()),
+    re_path(r'^zones/forward/(?P<name>(\d+/)?[^/]+)/nameservers$', views_zones.ForwardZoneNameServerDetail.as_view()),
+    path('zones/reverse/', views_zones.ReverseZoneList.as_view()),
+    re_path(r'^zones/reverse/(?P<name>(\d+/)?[^/]+)$', views_zones.ReverseZoneDetail.as_view()),
+    re_path(r'^zones/reverse/(?P<name>(\d+/)?[^/]+)/delegations/$', views_zones.ReverseZoneDelegationList.as_view()),
+    re_path(r'^zones/reverse/(?P<name>(\d+/)?[^/]+)/delegations/(?P<delegation>(.*))', views_zones.ReverseZoneDelegationDetail.as_view()),
+    re_path(r'^zones/reverse/(?P<name>(\d+/)?[^/]+)/nameservers$', views_zones.ReverseZoneNameServerDetail.as_view()),
     re_path(r'^zonefiles/(?P<name>(\d+/)?[^/]+)', views_zones.ZoneFileDetail.as_view()),
     path('permissions/netgroupregex/', views.NetGroupRegexPermissionList.as_view()),
     path('permissions/netgroupregex/<pk>', views.NetGroupRegexPermissionDetail.as_view()),
